Consolidate ingredient formatting from recipe and grocery list into one utility file IngredientFormatting. Comment out grocery quantity combination logic, instead just show as: 10 ml + 100 ml.
